Change all floats to doubles

+++ src/bbox.c
@@ -32,7 +32,7 @@ struct boundingBox *computeBoundingBox(struct poly *polys, int count) {
 		}
 	}
 	struct vector center = vectorWithPos(0.5 * (minPoint.x + maxPoint.x), 0.5 * (minPoint.y + maxPoint.y), 0.5 * (minPoint.z + maxPoint.z));
-	float maxDistance = 0.0;
+	double maxDistance = 0.0;
 	for (int i = 0; i < count; i++) {
 		for (int j = 0; j < 3; j++) {
 			struct vector fromCenter = subtractVectors(&vertexArray[polys[i].vertexIndex[j]], &center);
@@ -52,12 +52,12 @@ bool rayIntersectWithAABB(struct boundingBox *box, struct lightRay *ray, double 
 	dirfrac.y = 1.0f / ray->direction.y;
 	dirfrac.z = 1.0f / ray->direction.z;
 
-	float t1 = (box->start.x - ray->start.x)*dirfrac.x;
-	float t2 = (box->  end.x - ray->start.x)*dirfrac.x;
-	float t3 = (box->start.y - ray->start.y)*dirfrac.y;
-	float t4 = (box->  end.y - ray->start.y)*dirfrac.y;
-	float t5 = (box->start.z - ray->start.z)*dirfrac.z;
-	float t6 = (box->  end.z - ray->start.z)*dirfrac.z;
+	double t1 = (box->start.x - ray->start.x)*dirfrac.x;
+	double t2 = (box->  end.x - ray->start.x)*dirfrac.x;
+	double t3 = (box->start.y - ray->start.y)*dirfrac.y;
+	double t4 = (box->  end.y - ray->start.y)*dirfrac.y;
+	double t5 = (box->start.z - ray->start.z)*dirfrac.z;
+	double t6 = (box->  end.z - ray->start.z)*dirfrac.z;
 	
 	double tmin = max(max(min(t1, t2), min(t3, t4)), min(t5, t6));
 	double tmax = min(min(max(t1, t2), max(t3, t4)), max(t5, t6));
